refactor: move Car and Competition to domain package
