+++ src/test/java/TestCarRacingCompetition.java
@@ -1,10 +1,6 @@
-import CarRacing.*;
+import CarRacing.domain.Competition;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
